Clean up display of hidden posts
Visible posts are the normal state, and hidden posts are the exception - as such, it makes sense that the default styling is for visible posts, and anything for hidden posts is treated as an override. As such, have referred to hidden rather than visible, and have added a helper to make the code read more cleanly (if hidden makes more sense here than unless visible).

data-visible was on three seperate elements, which is overkill given that it's on the parent element.

Sass has been indented to help readbility.

Do not forget to include a space between the full colon and the style information in your sass.