Simplify autoexpand JS
This was pretty well written! But as always, I have some notes to help you out :)

The three argument version of $(document).on is useful when elements can be added or removed throughout the life-cycle of the page. However, because the textarea is always on the page, we can bind the event directly on the element itself

The module loader we have has a special way to bind JS directly to an element, called "apply", which you can check out here.

The code for setting the height of the element included a magic number, 17! What was this for, and was it dependant on a CSS line height? If so, this is fragile and can break. Instead what I do now is set the textarea to 0 pixels high, and then measure it's scrollHeight and update it straight away. This gives the same effect as your code, but is not dependant on any hard-coded numbers and is a bit simpler.

When targeting an element through it's classname, we preference the class with .js- . This ensures that classes used for style and classes used for JS never overlap.