Update permalink function
The permalink anchor allows us to link directly to a comment within the context of it's page, and scroll the browser down to the correct offset. I've included the link as a hidden link on the time_ago, so you can see it in action.

Every post has a unique url, in the format /post/:id, which then redirects to the correct page and scroll position within the thread. This is how we correctly show a new post when it is created.