Overhaul navigation layout and CSS structure
It was not my intention to make such a large change to the structure. While making a quick update to the user dropdown to use CSS instead of JS, I initially saw that the header and content area were not quite in alignment, and the page margins were far too large for intermediate page sizes. Diving into the CSS to investigate revealed quite a mess that had to be simplified before I could go any further. Some of the styling for the navigation and dropdown mayhave regressed slightly, but hopefully the improvements made to the foundations here will make it worth your while.

Please take some time to look over what I've done, but here are the main points:

- .hdrNav and .hdrToolbar are renamed .navigation and .toolbar - we don't abbreviate in class names unless absolutely neccesary, and it's understood that these are both elements that would only live inside the header.
- User dropdown is now CSS like the post moderation toolbar, removing the need for any JS
- There were two almost identical classes in breakpoint.css, .constrained and .group. Constrained set the page margins unusually high, and group was redundant (and later removed completely when simplifying the markup for the navigation)
- Get support was moved from global to footer
- Styles relating to .header were moved into the header scss file, which feels much more logical
- content.scss only contained three classes, all of which are mostly including mixins that are used no-where else. There has been a few times recently where I've refactored out a mixin that would only ever be included once (the dark and light strips are a good example of this). I inlined and simplified the CSS, then placed all three styles into global.scss, given that these are global styles
- Header.scss has been cleaned up a lot, and lots of floats taken out. I'll need to go into more detail about using floats vs inline block vs absolute/relative postitioning, but hopefully this simpler CSS and HTML will make debugging and breakpoint styling easier