Modal form cleanup
I've cleaned up several aspects of the popup styling. These ideas should be applied across the rest of the app before proceeding.

A consensus needs to be made in regards to the vendor prefixes in CSS. There are currently some inlined vendor prefixes, some done with the specific cross-browser sass functions, and some properties that have not been vendor prefixed. A choice needs to be made and applied uniformly across the entire codebase. My suggestion that it's either the SASS function, or it's done with an auto-prefixing gem. Maintaining vendor-prefixes in a non-centralised way is prohibitive.

Reverted the dropdown back to the original functionality I had created. Using JS to help style the display of the selected value is fine, but overriding the default form control is a no-no. It removes accessibility and keyboard support provided by default by the browser, and introduces issues when attempting to make the dropdown behave uniformly across all platforms and browsers.

Moved styles from old dropdown to new one where possible.

There were commented out lines in JS and slim that shouldn't be there - commented out lines should never be committed.

Clean up modal.scss. The styles were not scoped to be specific to just the modal layout, so I've added a modalContainer class and re-scoped the CSS where possible. The .constrained has also been removed in anticipation of the popup requiring different styling to the main page.

Removed all the <strong>s inside .label__title. The title is always to be displayed in bold, so there is no need to introduce extra markup.

The toggle has been reworked to remove the need to hardcode dimensions based on the text label sizes, and redundant markup and styles removed. Hard-coding dimensions introduces extra work every time the widget is used, and needs to be re-done for differing fontsizes (such as in different breakpoints), and will be wrong if font-loading fails.

@unselectable should be used instead of user-select: none;

Transitions of 400ms seem too long. Check with the designer, I've changed them to 200ms on the toggle for the time being.