Clean-up post content cropping
The JS was inextricably linked to the current HTML structure, and changing anything about the structure would break the JS, given then number of chained parent() and find() calls. This can all be simplified to simply toggling two classes on the parent element, and handling the visual changes in CSS. This also removes the need to hard-code text values in the JS, which can be more easily set in the markup.

The transitions and gradient codes will need to be updated to reflect the decision made with vendor-prefixing.