Show modal content as a popup, and floating notifications
I've used the edit button on the topic thread as an example of how to launch the modal content in a popup window, this can now be rolled out across the app by adding data-modal=true to the relevant links.

The content is fetched and inserted into .modalContainer. The close button needs to be styled, as does the loading message while the content is being fetched; both have been noted for you in _modal.scss.

I have also added success and error notifications to the main app. You can see an example by saving a change to a topic thread page. Again these have styling that need to be set properly, notes are in global.scss. There's some JS to hide them after 5 seconds, you can change this value depending on the feedback from the designer.