Simplify sidebar heading
h1 Should realistically only exist once per a page, for the overall page heading. As such it does not make sense to use it as a subheader in a sidebar. Also the strong was redundant as the bold styling can be applied directly to the heading. Please remove any instances of nested strong tags and incorrect h1 tags across the app.