Simplify markup for sidebar buttons, requires attention
Placing a second element inside of the link here is redundant, and also places a block-level element (the div) inside of an inline one (the anchor) which is a no-no. Moving the class to the anchor makes more sense here, please roll this out across anywhere else in the sidebar and elsewhere.