Beginnings of error pages
Error pages need to be marked up, styled, and any text content added. Because error pages need to work when the app doesn't, they can not contain any dynamic templating or CSS, or reference the compiled CSS. Because of this, I strongly suggest a very simple design with little to no images or custom typefaces (if you need to add assets, they need to go in /public alongside the placeholder error.css file I've added).

The error pages are 404, 403, 500, and 502. To view any of them, you can go direclty to them e.g. pete-app.dev/404.html