Massive reorganisation and tidy up of thread scss code.
JB: Please review and check this file, let me know if it's too anything and I'll re-work it. Once you're cool with this, I'll be going through the rest of my scss files and applying the same/similar coding style to tidy/re-name everything. So lemme know what you think!