Add ability to quote posts
Clicking the "quote" button copies the posts id to a hidden field "quoted_post_id", which is used to save a reference to the quoted post in the post being made.
The ID is validated to make sure it is a post that exists in the same thread, and that hasn't been hidden by a moderator.
A display above the post entry box indicates a brief description of the post being quoted, along with a button to remove the quotation.
The visibility of the entire quoted section is controlled with a class "active".
Posts render quoted posts with the same partial, but with an added parameter "being_quoted", to stop recursion and to make sure controls for quoting, moderation etc don't appear within the quoted post. The entire quoted post is within .post__quoted, which can be targeted to provide an alternate styling.