Allow schedulfed_for to be edited as two seperate fields
The date and time of a chat_sessions's scheduled_for can be accessed through scheduled_for_date and scheduled_for_time. The two values that are set in the form will then be combined back into one scheduled_for datetime when saving.

The values that are being set initially in the form are correct, but the datepicker is not highlighting the selected day and the datepicker is reverting to the current time when clicking on it, these both need to be addressed by MC