Make notification details editable on users
The general toggle for notifications has been removed, given that there are now specific checkboxes for each type of notification.

I had to massacre the form to get it working; having a mix of click-to-edit and checkboxes without submit buttons makes it very messy; so for the sake of getting the functionality there I have stripped it back to text boxes and a single submit form. MC to style as required, now that functionality is restored (also removed a modal class, no idea what that was doing in the page...)